windows: call WaitCommEvent after processing monome input buffer
fixes #68. if WaitCommEvent is called on every loop iteration, it
may result in the monome input event ending up in a non-signalled
state when there's OSC traffic being received in the background.

if SetCommMask/WaitCommEvent are both called after monome input
buffer is fully processed (and just before the loop), the issue
is no longer happening.